achievement-2/Exercise_2.2/A2_Recipe_App/src/requirements/base.txt
Django~=5.0
django-environ==0.11.2
